+++ daemon/src/auth.rs
@@ -1,4 +1,5 @@
 use hex::FromHexError;
+use rocket::http::Header;
 use rocket::http::Status;
 use rocket::outcome::try_outcome;
 use rocket::outcome::IntoOutcome;
@@ -14,7 +15,7 @@ use std::str::FromStr;
 /// A request guard that can be included in handler definitions to enforce authentication.
 pub struct Authenticated {}
 
-pub const MAKER_USERNAME: &str = "maker";
+pub const USERNAME: &str = "itchysats";
 
 #[derive(Debug)]
 pub enum Error {
@@ -28,17 +29,17 @@ pub enum Error {
 }
 
 #[derive(PartialEq)]
-pub struct Password([u8; 32]);
+pub struct Password(String);
 
 impl From<[u8; 32]> for Password {
     fn from(bytes: [u8; 32]) -> Self {
-        Self(bytes)
+        Self(hex::encode(bytes))
     }
 }
 
 impl fmt::Display for Password {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0))
+        self.0.fmt(f)
     }
 }
 
@@ -46,10 +47,7 @@ impl FromStr for Password {
     type Err = FromHexError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut bytes = [0u8; 32];
-        hex::decode_to_slice(s, &mut bytes)?;
-
-        Ok(Self(bytes))
+        Ok(Self(s.to_string()))
     }
 }
 
@@ -68,7 +66,7 @@ impl<'r> FromRequest<'r> for Authenticated {
             .await
             .map_failure(|(status, _)| (status, Error::MissingPassword)));
 
-        if basic_auth.username != MAKER_USERNAME {
+        if basic_auth.username != USERNAME {
             return Outcome::Failure((
                 Status::Unauthorized,
                 Error::UnknownUser(basic_auth.username),
@@ -86,5 +84,76 @@ impl<'r> FromRequest<'r> for Authenticated {
         }
 
         Outcome::Success(Authenticated {})
+    }
+}
+
+/// A "catcher" for all 401 responses, triggers the browser's basic auth implementation.
+#[rocket::catch(401)]
+pub fn unauthorized() -> PromptAuthentication {
+    PromptAuthentication {
+        inner: (),
+        www_authenticate: Header::new("WWW-Authenticate", r#"Basic charset="UTF-8"#),
+    }
+}
+
+/// A rocket responder that prompts the user to sign in to access the API.
+#[derive(rocket::Responder)]
+#[response(status = 401)]
+pub struct PromptAuthentication {
+    inner: (),
+    www_authenticate: Header<'static>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use rocket::http::Status;
+    use rocket::local::blocking::Client;
+    use rocket::Build;
+    use rocket::Rocket;
+
+    #[test]
+    fn routes_are_password_protected() {
+        let client = Client::tracked(rocket()).unwrap();
+
+        let response = client.get("/protected").dispatch();
+
+        assert_eq!(response.status(), Status::Unauthorized);
+        assert_eq!(
+            response.headers().get_one("WWW-Authenticate"),
+            Some(r#"Basic charset="UTF-8"#)
+        );
+    }
+
+    #[test]
+    fn correct_password_grants_access() {
+        let client = Client::tracked(rocket()).unwrap();
+
+        let response = client.get("/protected").header(auth_header()).dispatch();
+
+        assert_eq!(response.status(), Status::Ok);
+    }
+
+    #[rocket::get("/protected")]
+    async fn protected(_auth: Authenticated) {}
+
+    /// Constructs a Rocket instance for testing.
+    fn rocket() -> Rocket<Build> {
+        rocket::build()
+            .manage(Password::from(*b"Now I'm feelin' so fly like a G6"))
+            .mount("/", rocket::routes![protected])
+            .register("/", rocket::catchers![unauthorized])
+    }
+
+    /// Creates an "Authorization" header that matches the password above,
+    /// in particular it has been created through:
+    /// ```
+    /// base64(itchysats:hex("Now I'm feelin' so fly like a G6"))
+    /// ```
+    fn auth_header() -> Header<'static> {
+        Header::new(
+            "Authorization",
+            "Basic aXRjaHlzYXRzOjRlNmY3NzIwNDkyNzZkMjA2NjY1NjU2YzY5NmUyNzIwNzM2ZjIwNjY2Yzc5MjA2YzY5NmI2NTIwNjEyMDQ3MzY=",
+        )
     }
 }